More work towards a stable persistent connection implementation.
The demos  folder "Setup" is  replaced by a "Factory"  component which
lives in  the main library.  Some  of the demos are  refactored to use
the Factory class.
Improved the channel flow API level support in the main Channel class,
including new methods to toggle flow.
Tweaks in the StreamSocket implementation to expose more of the
underlying connection metadata through the API.
Added  a   "suspend  flow"  feature   to  PChannel,  this   means  the
implementation can  automatically set  the channel to  flow=yes during
rehydration and flow=no during  serialize.  Most of the implementation
of this is in PConnection and Channel classes.