Part way through lost messages debugging.
Look for BINGOROONIE - this displays the message that will be lost.

There's a  mechanism to persist the  ftell() result to  try and detect
lost  messages, this  won't work  and needs  to be  removed  ($z[3] in
PConnection->serialize())

Looks like the best thing to do is keep the nbRead method and poke any
content which will be lost in to Connection->$readSrc so that at least
it's not lost.